Update byte to gb function so rounding is the same between filter and aggregate file size methods, to keep counts the same

+++ utilities.py
@@ -7,7 +7,7 @@ import collections
 
 def bytes_to_gb(bytes):
     try:
-        return float(bytes) / (1024**3.0)
+        return int( float(bytes) / (1024**3.0) )
     except ValueError:
         return bytes
